Fix draft runtime capabilities
