+++ packages/server-ai/src/xpert/xpert.module.ts
@@ -28,6 +28,8 @@ import { AgentViewHostDefinition } from '../view-extension/hosts/agent-view-host
 import { SkillPackageModule } from '../skill-package'
 import { XpertTemplateModule } from '../xpert-template/xpert-template.module'
 import { PromptWorkflowModule } from '../prompt-workflow'
+import { RuntimeCapabilitiesService } from '../ai/runtime-capabilities.service'
+import { RuntimeCommandService } from '../ai/runtime-command.service'
 
 @Module({
     imports: [
@@ -64,6 +66,8 @@ import { PromptWorkflowModule } from '../prompt-workflow'
         PublishedXpertAccessService,
         XpertAuthoringService,
         XpertAuthoringMiddleware,
+        RuntimeCommandService,
+        RuntimeCapabilitiesService,
         ...CommandHandlers,
         ...QueryHandlers
     ],
